Lock based migration operations
Takes a named lock prior to executing the migration operation.
The lock is named with the operation executed and the id of the
source VM. This lock can then be used by other operations to avoid
operation conflicts on shared resources.

Example use:
If a migration operation uses this decorator, it will block until
the named lock is free. This is used to protect concurrent migration
operations of the same source VM.